pyproject.toml
[project]
name = "bt-python-sdk"
version = "1.0.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.1",
    "pytest-randomly>=3.12.0",
    "pytest-timeout>=2.1.0",
    "pytest-xdist>=3.3.1",
    "ruff>=0.11.9",
]
